edited logging
